+++ packages/google_sign_in/google_sign_in_android/CHANGELOG.md
@@ -1,6 +1,10 @@
-## 6.1.11
+## 6.1.12
 
 * Adds a namespace for compatibility with AGP 8.0.
+
+## 6.1.11
+
+* Fixes Java warnings.
 
 ## 6.1.10