+++ packages/local_auth/local_auth_android/CHANGELOG.md
@@ -1,6 +1,10 @@
-## 1.0.24
+## 1.0.25
 
 * Adds a namespace for compatibility with AGP 8.0.
+
+## 1.0.24
+
+* Fixes `getEnrolledBiometrics` return value handling.
 
 ## 1.0.23